Full project file structure refactor
