+++ src/components/ui/SalesBanner.jsx
