+++ src/main.jsx
@@ -3,10 +3,10 @@ import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App.jsx";
 import { BrowserRouter } from "react-router-dom";
-import { EmployeeProvider } from "./context/EmployeeContext.jsx";
-import { CategoryProvider } from "./context/CategoryContext.jsx";
-import { ProductProvider } from "./context/ProductContext.jsx";
-import { CartProvider } from "./context/CartContext.jsx";
+import { EmployeeProvider } from "@context/EmployeeContext.jsx";
+import { CategoryProvider } from "@context/CategoryContext.jsx";
+import { ProductProvider } from "@context/ProductContext.jsx";
+import { CartProvider } from "@context/CartContext.jsx";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
