+++ src/pages/admin/Category.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
-import { useCategory } from "../context/CategoryContext";
-import CategoryModalUpsert from "./CategoryModalUpsert";
-import CategoryModalDelete from "./CategoryModalDelete";
+import { useCategory } from "@context/CategoryContext";
+import CategoryModalUpsert from "@modal/CategoryModalUpsert";
+import CategoryModalDelete from "@modal/CategoryModalDelete";
 
 const Category = () => {
   const { categories, deleteCategory } = useCategory();
